Internal file reorganization
